Layers and overlays
Looks like imgui layer works, it renders behind the character which is interesting and i don't know for sure why is that, maybe because of the order of rendering in the main loop because i render the layer stack first which at the moment has only imgui and only then the model.

Looks nice

+++ include/core/renderer/camera.h
@@ -39,7 +39,7 @@ namespace kogayonon
     const glm::mat4& getViewMatrix()const;
 
     void processMouseMoved(float x, float y, bool constrain_pitch = true);
-    void processKeyboard(GLFWwindow* window, float delta_time);
+    void processKeyboard(GLFWwindow* window, double delta_time);
 
     void updateCameraVectors();
